make phenomodeladapter an indipendent class
